chore: remove mini:v8
