+++ Cargo.toml
@@ -49,7 +49,6 @@ async-graphql = { version = "7.0.1", features = [
 async-graphql-value = "7.0.1"
 indexmap = "2.2"
 once_cell = "1.19.0"
-mini-v8 = { version = "0.4.1", optional = true }
 clap = { version = "4.4.18", features = ["derive"] }
 colored = "2"
 regex = "1.10.3"
@@ -81,8 +80,8 @@ prost = "0.12.3"
 update-informer = { version = "1.1.0", default-features = false, features = ["github", "reqwest"], optional = true }
 lazy_static = "1.4.0"
 which = { version = "6.0.0", optional = true }
-deno_core = "0.257.0"
-serde_v8 = "0.167.0"
+deno_core = { version = "0.257.0", optional = true }
+serde_v8 = { version= "0.167.0", optional = true }
 
 [dev-dependencies]
 criterion = "0.5.1"
@@ -95,8 +94,8 @@ insta = { version = "1.34.0", features = ["json"] }
 [features]
 
 # Feature Flag to enable V8.
-# V8 currently is not support on all platforms so we control it via this feature flag.
-js = ["dep:mini-v8"]
+# V8 currently is not support on all platforms so, we control it via this feature flag.
+js = ["dep:deno_core", "dep:serde_v8"]
 
 # Feature Flag to core CLI features.
 # This is created to control what we expose for WASM.
