+++ src/cli/javascript/mod.rs
@@ -1,15 +1,10 @@
-mod http_filter_mv8;
-mod runtime_mv8;
-mod serde_mv8;
-mod shim;
 pub use std::sync::Arc;
-mod channel_mv8;
-mod deno_channel;
-mod deno_http_filter;
-mod deno_runtime;
+mod channel;
+mod http_filter;
+mod runtime;
 
-pub use deno_http_filter::HttpFilter;
-pub use deno_runtime::Runtime;
+pub use http_filter::HttpFilter;
+pub use runtime::Runtime;
 
 use crate::{blueprint, HttpIO};
